Fixes a few sonarqube violations.

+++ org.sat4j.core/src/main/java/org/sat4j/minisat/core/SizeComparator.java
@@ -51,10 +51,10 @@ public class SizeComparator implements Comparator<Constr>, Serializable {
      * @see java.util.Comparator#compare(java.lang.Object, java.lang.Object)
      */
     public int compare(Constr c1, Constr c2) {
-        long delta = c1.size() - c2.size();
+        int delta = c1.size() - c2.size();
         if (delta == 0) {
             return (int) Math.round(c2.getActivity() - c1.getActivity());
         }
-        return (int) delta;
+        return delta;
     }
 }
